Added browserview to get custom_css styles for tinymce.
