+++ plonetheme/onegov/configure.zcml
@@ -19,6 +19,7 @@
 
   <include package=".browser" />
   <include package=".portlets" />
+  <include package=".upgrades" />
   <include package=".viewlets" />
   <include zcml:condition="installed ftw.lawgiver" file="lawgiver.zcml" />
