+++ plonetheme/onegov/profiles/default/metadata.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <metadata>
-  <version>1000</version>
+  <version>1001</version>
   <dependencies>
     <dependency>profile-plone.app.theming:default</dependency>
     <dependency>profile-collective.mtrsetup:default</dependency>
