+++ plonetheme/onegov/profiles/default/tinymce.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<object>
+  <layout>
+    <content_css value="customstyles_css"/>
+  </layout>
+</object>
