+++ plonetheme/onegov/upgrades/configure.zcml
@@ -1,0 +1,20 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
+    xmlns:upgrade-step="http://namespaces.zope.org/ftw.upgrade"
+    xmlns:zcml="http://namespaces.zope.org/zcml"
+    i18n_domain="plonetheme.onegov">
+
+  <include package="ftw.upgrade" file="meta.zcml" />
+
+  <!-- 1000 -> 1001 -->
+  <upgrade-step:importProfile
+      title="Update tinymce css"
+      profile="plonetheme.onegov:default"
+      source="1000"
+      destination="1001"
+      directory="profiles/1001"
+      />
+
+
+</configure>
