+++ setup.py
@@ -39,6 +39,7 @@ setup(name='plonetheme.onegov',
         'collective.mtrsetup',
         'ftw.mobilenavigation',
         'ftw.slider',
+        'ftw.upgrade',
         'setuptools',
         'Missing',
         'PyScss',
